+++ R/anti_sample.R
@@ -1,4 +1,9 @@
-## get request with sampling
+#' Get a GA request with anti-sampling
+#' 
+#' Calculates multiple API calls to avoid sampling
+#' 
+#' @inheritParams make_ga_4_req
+#' @keywords internal
 anti_sample <- function(viewId,
                         date_range,
                         metrics,
@@ -29,13 +34,13 @@ anti_sample <- function(viewId,
                                   metricFormat      = metricFormat,
                                   samplingLevel     = "LARGE",
                                   histogramBuckets  = histogramBuckets)
-  
-  read_counts <- as.integer(attr(test_call,"samplesReadCounts")[[1]])
-  space_size <- as.integer(attr(test_call, "samplingSpaceSizes")[[1]])
-  samplingPer <- get_samplePercent(read_counts, space_size)
-  rowCount <- as.integer(attr(test_call, "rowCount")[[1]])
   ## adds 10% to read_counts as its flakey
-  m_error <- 0.9 
+  read_counts <- as.integer(attr(test_call,"samplesReadCounts")[[1]]) * 0.9
+  space_size  <- as.integer(attr(test_call, "samplingSpaceSizes")[[1]])
+  samplingPer <- get_samplePercent(read_counts, space_size)
+  
+  ## add 20% to rowCount as its flakey
+  rowCount <- as.integer(attr(test_call, "rowCount")[[1]]) * 1.2
   
   if(identical(samplingPer, numeric(0))){
     message("No sampling found, returning call")
@@ -49,15 +54,12 @@ anti_sample <- function(viewId,
                                          metrics = "sessions",
                                          dimensions = "date")
   explore_sessions$cumulative <- cumsum(explore_sessions$sessions)
-  explore_sessions$sample_bucket <- as.factor((explore_sessions$cumulative %/% (read_counts*m_error) + 1))
+  explore_sessions$sample_bucket <- as.factor((explore_sessions$cumulative %/% read_counts) + 1)
   splits <- split(explore_sessions, explore_sessions[["sample_bucket"]])
   
   new_date_ranges <- lapply(splits, function(x) {list(start_date = min(x$date), 
                                                      end_date = max(x$date),
                                                      range_date = nrow(x))})
-  
-  ## if any new_date_ranges range_date is 1 then possibily will still sample.
-  ## split by hour/day
   
   ## send to fetch
   myMessage("Calculated [", length(new_date_ranges), "] batches are needed to download [", rowCount,"] rows unsampled.", level = 3)
@@ -81,12 +83,36 @@ anti_sample <- function(viewId,
                               metricFormat      = metricFormat,
                               samplingLevel     = "LARGE",
                               histogramBuckets  = histogramBuckets)
+    
     read_counts2 <- as.integer(attr(out,"samplesReadCounts")[[1]])
-    space_size2 <- as.integer(attr(out, "samplingSpaceSizes")[[1]])
-    samplingPer <- get_samplePercent(read_counts2, space_size2)
+    space_size2  <- as.integer(attr(out, "samplingSpaceSizes")[[1]])
+    samplingPer  <- get_samplePercent(read_counts2, space_size2)
+    
     if(!identical(samplingPer, numeric(0))){
       myMessage("Anti-sampling failed", level = 3)
-      did_it_work <<- FALSE
+      if(x$range_date == 1){
+        ## if any new_date_ranges range_date is 1 then possibily will still sample.
+        myMessage("Attempting hourly anti-sampling...", level = 3)
+        out <- hourly_anti_sample(viewId            = viewId,
+                                  the_day           = x$start_date,
+                                  metrics           = metrics,
+                                  dimensions        = dimensions,
+                                  dim_filters       = dim_filters,
+                                  met_filters       = met_filters,
+                                  filtersExpression = filtersExpression,
+                                  order             = order,
+                                  segments          = segments,
+                                  pivots            = pivots,
+                                  cohorts           = cohorts,
+                                  max               = attr(out, "rowCount"),
+                                  metricFormat      = metricFormat,
+                                  samplingLevel     = "LARGE",
+                                  histogramBuckets  = histogramBuckets,
+                                  read_counts = read_counts)
+      } else {
+        did_it_work <<- FALSE
+      }
+
     }
     out
   })
@@ -100,4 +126,81 @@ anti_sample <- function(viewId,
   myMessage("Finished unsampled data request, total rows [", nrow(out),"]", level = 3)
   if(did_it_work) myMessage("Successfully avoided sampling", level = 3)
   out
+}
+
+#' hourly get request with anti-sampling
+#' @keywords internal
+hourly_anti_sample <- function(viewId,
+                               the_day,
+                               metrics,
+                               dimensions,
+                               dim_filters,
+                               met_filters,
+                               filtersExpression,
+                               order,
+                               segments,
+                               max,
+                               pivots,
+                               cohorts,
+                               metricFormat,
+                               histogramBuckets,
+                               read_counts){
+  
+  ## get session distribution per hour
+  ## sampling
+  myMessage("Finding number of hourly sessions for anti-sample calculations...", level = 3)
+  explore_sessions <- google_analytics_4(viewId = viewId,
+                                         date_range = c(the_day, the_day),
+                                         metrics = "sessions",
+                                         dimensions = "hour")
+  
+  ## work out batches
+  explore_sessions$cumulative <- cumsum(explore_sessions$sessions)
+  explore_sessions$sample_bucket <- as.factor((explore_sessions$cumulative %/% read_counts) + 1)
+  splits <- split(explore_sessions, explore_sessions[["sample_bucket"]])
+  
+  new_hour_ranges <- lapply(splits, function(x) {list(hours = x$hour,
+                                                      range_date = nrow(x))})
+
+  
+  ## do calls
+  unsampled_list <- lapply(new_hour_ranges, function(x){
+
+    myMessage("Anti-sample call covering ", x$range_date, " hours: ", paste(x$hours, collapse = " "), 
+              level = 3)
+    
+    hour_filter <- filter_clause_ga4(list(dim_filter("hour", "IN_LIST", x$hours)))
+    
+    if(!is.null(dim_filters)){
+      stop("Can't use dim_filters argument with hourly anti-sampling, sorry!  
+           Can you use filtersExpression instead?")
+    }
+    
+    out <- google_analytics_4(viewId            = viewId,
+                              date_range        = c(the_day, the_day),
+                              metrics           = metrics,
+                              dimensions        = dimensions,
+                              dim_filters       = hour_filter,
+                              met_filters       = met_filters,
+                              filtersExpression = filtersExpression,
+                              order             = order,
+                              segments          = segments,
+                              pivots            = pivots,
+                              cohorts           = cohorts,
+                              max               = max,
+                              metricFormat      = metricFormat,
+                              samplingLevel     = "LARGE",
+                              histogramBuckets  = histogramBuckets)
+    
+    read_counts3 <- as.integer(attr(out,"samplesReadCounts")[[1]])
+    space_size3  <- as.integer(attr(out, "samplingSpaceSizes")[[1]])
+    samplingPer  <- get_samplePercent(read_counts3, space_size3)
+    
+    if(!identical(samplingPer, numeric(0))){
+      myMessage("Hourly anti-sampling failed. Holy-moly, you should try GA 360 and BigQuery.", level = 3)
+    }
+    out
+  })
+  ## output rbind
+  Reduce(rbind, unsampled_list)
 }